refactor(core-sdk): models/blocks package; move vit_ops to blocks/vit
blocks/ hosts composition helpers (pipeline-fragment builders, no
@register_model). vit_ops.py moves byte-identical to blocks/vit.py;
a compatibility shim stays at models/vit_ops.py for external imports,
qwen3vl imports from the new location, and the app side gains the
blocks package facade + blocks.vit leaf wrapper with a
CORE_SDK_LEAF_MODULES contract entry.
