+++ aic-core/src/aiconfigurator_core/sdk/models/blocks/__init__.py
@@ -1,0 +1,6 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+from .vit import build_encoder_ops
+
+__all__ = ["build_encoder_ops"]
