+++ src/aiconfigurator/sdk/models/blocks/vit.py
@@ -1,0 +1,8 @@
+# SPDX-FileCopyrightText: Copyright (c) 2025-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Compatibility alias for aiconfigurator_core.sdk.models.blocks.vit."""
+
+from aiconfigurator.sdk._compat import alias_module as _alias_module
+
+_alias_module(__name__, "aiconfigurator_core.sdk.models.blocks.vit")
